+++ src/components/Inputs/Slider/types.ts
@@ -1,10 +1,6 @@
-import * as GlobalInterfaces from "../../global-types";
+import { InputProps } from "../_inputsComponents/input-types";
 
-export type ValuePreviewType =
-  | "top-dynamic"
-  | "bottom-dynamic"
-  | "top-static"
-  | "bottom-static";
+export type ValuePreviewType = "top-dynamic" | "bottom-dynamic" | "top-static" | "bottom-static";
 
 type SliderClassNames = {
   container?: string;
@@ -14,23 +10,15 @@ type SliderClassNames = {
   valuePreview?: string;
 };
 
-export type SliderProps = {
+export type SliderProps = InputProps & {
   value?: number;
   min: number;
   max: number;
   step?: number;
   initialValue?: number;
   onChange?: (event: React.ChangeEvent<HTMLInputElement>) => void;
-  name?: string;
-  labelWidth?: GlobalInterfaces.LabelPercentageWidth;
-  floatingInputWidth?: GlobalInterfaces.FloatingInputWidth;
-  size?: `${GlobalInterfaces.ComponentSize}`;
-  label?: string;
-  labelType?: `${GlobalInterfaces.InputLabel}`;
   hideValuePreview?: boolean;
-  disabled?: boolean;
-  readOnly?: boolean;
   valuePreviewType?: ValuePreviewType;
-  noBottomMargin?: boolean;
   classNamesObj?: SliderClassNames;
+  error?: string;
 };